+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.6.7
 
+- Remove outdated permissions browser bug warning
+
 ## 0.6.6
 
 - Write reported item to quick message body